+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Versioning since version 1.0.0.
 
 ### Fixed
 
+- Fix for no longer removing "bolded" images
+
 ## [3.25.0] - 2026-01-20
 
 ### Added